+++ NEWS.md
@@ -138,6 +138,12 @@
   space. `conn_meow()` is like `meow()` but takes a connection as
   first argument.
 
+* New `box` class. Its purpose is similar to the `AsIs` class from
+  `base::I()`, i.e. it protects a value temporarily. However it does
+  so by wrapping the value in a scalar list. Use `box()` to create a
+  boxed value, `is_box()` to test for a boxed value, and `unbox()` to
+  unbox it. `box()` and `is_box()` accept optional subclass.
+
 
 ## Breaking changes